[generator] remove extended scalars (#1171)
Extended scalars were deprecated in `graphql-java` v15 in favor of separate `graphql-java-extended-scalars` project. Extended scalars are problematic as clients have to explicitly know how to handle them and what they represent. Due to the all the drawbacks, instead of automatically providing support for deprecated functionality, users should explicitly opt-in into this functionality by using custom hooks.

+++ examples/server/spring-server/src/main/kotlin/com/expediagroup/graphql/examples/server/spring/subscriptions/SimpleSubscription.kt
@@ -40,7 +40,7 @@ class SimpleSubscription : Subscription {
     fun singleValueSubscription(): Flux<Int> = Flux.just(1)
 
     @GraphQLDescription("Returns a random number every second")
-    fun counter(limit: Long? = null): Flux<Int> {
+    fun counter(limit: Int? = null): Flux<Int> {
         val flux = Flux.interval(Duration.ofSeconds(1)).map {
             val value = Random.nextInt()
             logger.info("Returning $value from counter")
@@ -48,7 +48,7 @@ class SimpleSubscription : Subscription {
         }
 
         return if (limit != null) {
-            flux.take(limit)
+            flux.take(limit.toLong())
         } else {
             flux
         }
